Move new async-replication tests to old file
