--- tests/spread/test_async_replication_new.py/task.yaml
@@ -1,9 +1,0 @@
-summary: test_async_replication_new.py
-environment:
-  TEST_MODULE: high_availability/test_async_replication_new.py
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
-variants:
-  - -juju29
